+++ modules/server/src/main/ts/bedrock/cli/Clis.ts
@@ -51,6 +51,7 @@ export const forAuto = (directories: Directories, argv: string[] = process.argv)
     ClOptions.wipeBrowserCache,
     ClOptions.remote,
     ClOptions.webdriverPort,
+    ClOptions.pageHost,
     ClOptions.sishDomain,
     ClOptions.username,
     ClOptions.accesskey,